requirements.txt
chromedriver-binary==87.0.4280.20.0
python-dotenv==0.15.0
selenium==3.141.0
urllib3==1.26.2
